% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdiagnose.R
\name{cdiagnose}
\alias{cdiagnose}
\title{Graphically diagnose model residuals ("classic" version with ggplot2 graphics).}
\usage{
cdiagnose(model, se = FALSE, alpha = 1)
}
\arguments{
\item{model}{An lm or glm object.}

\item{se}{Boolean. To overly standard errors.}

\item{alpha}{Integer, [0, 1]. Points are more transparent the closer they are to 0.}
}
\value{
2x2 charts that closely resemble the output to plot(model.lm) with ggplot2 graphics.
}
\description{
Graphically diagnose model residuals ("classic" version with ggplot2 graphics).
}
\examples{
model.lm <- lm(data = mtcars, formula = mpg ~ wt + gear)
cdiagnose(model.lm)
}
\seealso{
\url{https://github.com/robertschnitman/diagnoser} and
Raju Rimal (inspiration for Residuals vs. Leverage): \url{https://rpubs.com/therimalaya/43190}
}
