% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleancorpus.R
\name{cleancorpus}
\alias{cleancorpus}
\title{Clean corpus by removing punctuation, stripping whitespace, and others.}
\usage{
cleancorpus(corpus, stopwords)
}
\arguments{
\item{corpus}{A corpus document.}

\item{stopwords}{Specification for stopwords.}
}
\value{
A corpus.
}
\description{
Clean corpus by removing punctuation, stripping whitespace, and others.
}
\examples{
library(tm)
cleancorpus(my_corpus, 'english')
}
\seealso{
\url{https://github.com/robertschnitman/schnitr}
}
