% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnoseR.R
\name{comp_alg}
\alias{comp_alg}
\title{Compare Algoritmos de Amprendizado de Máquina.}
\usage{
comp_alg(
  data,
  target,
  list_alg = c("rpart", "nnet", "svmLinear", "rf", "LogitBoost", "knn"),
  train_val = 0.75,
  seed = 123,
  number = 5,
  repeats = 10,
  verbose = TRUE
)
}
\arguments{
\item{data}{The dataframe containing your data.}

\item{target}{A string with the name of the column you want to predict.}

\item{list_alg}{A character vector of the machine learning algorithms you want to compare.}

\item{train_val}{The proportion of the data to be used for training.}

\item{seed}{A random seed for reproducibility.}

\item{number}{The number of folds for repeated cross-validation.}

\item{repeats}{The number of times to repeat the cross-validation.}

\item{verbose}{If `TRUE`, it prints the confusion matrix for each algorithm during execution.}
}
\value{
A list object of class `diagnoseR_result` containing models, evaluations, and performance metrics.
}
\description{
Esta função compara multiplos algoritmos de aprendizado de máquina.
}
