% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{diagnose.lm}
\alias{diagnose.lm}
\title{Diagnose a lm model}
\usage{
\method{diagnose}{lm}(model)
}
\arguments{
\item{model}{lm object}
}
\description{
Diagnose a lm model
}
\note{
To diagnose the lm model
}
