% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{diagnose.lme}
\alias{diagnose.lme}
\title{diagnose.lme}
\usage{
\method{diagnose}{lme}(model, ...)
}
\arguments{
\item{model}{\code{\link[nlme]{lme}} object.}

\item{...}{Currently not to be used.}
}
\value{
An object of \code{tbl_df_diag}.
}
\description{
To generate a diagnosis of a \code{lme} model.
}
\note{
To diagnose the \code{\link[nlme]{lme}} model.
}
\examples{
library(nlme)
lme_orthodont <- lme(distance ~ age + Sex, Orthodont)
(diag_orthodont <- diagnose(lme_orthodont))
}
