% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnoser.R
\docType{package}
\name{diagnoser}
\alias{diagnoser}
\title{diagnoser: A package for diagnosis of statistical models}
\description{
The diagnoser package provides diagnosis set of statistical models. It also provides the plot of diagnosis.
}
