% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnosis.R
\docType{data}
\name{GeomPointLoess}
\alias{GeomPointLoess}
\title{GeomPointLoess}
\format{
An object of class \code{GeomPointLoess} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
GeomPointLoess
}
\description{
GeomPointLoess
}
\keyword{datasets}
