% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnoseR.R
\name{get_var_importance}
\alias{get_var_importance}
\title{Get Variable Importance from Trained Models}
\usage{
get_var_importance(diagnoseR_result)
}
\arguments{
\item{diagnoseR_result}{A list object of class `diagnoseR_result` from `comp_alg`.}
}
\value{
A list where each element is a data frame of variable importances for an algorithm,
sorted from most to least important.
}
\description{
This function extracts and summarizes variable importance from the models
trained by `comp_alg`.
}
