% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdiagnose.R
\name{ggdiagnose}
\alias{ggdiagnose}
\title{Graphically diagnose model residuals (ggplot2 version).}
\usage{
ggdiagnose(model, bins = 30, se = TRUE, freqpct = FALSE, alpha = 1)
}
\arguments{
\item{model}{An lm or glm object.}

\item{bins}{Number of bins to specify for histograms.}

\item{se}{Boolean. For overlaying shaded standard errors.}

\item{freqpct}{Boolean.}

\item{alpha}{Integer, [0, 1]. Points are more transparent the closer they are to 0. Only applies to scatter plots.}
}
\value{
2x2 charts similar to plot(model): 2 scatter plots and 2 histograms.
}
\description{
Graphically diagnose model residuals (ggplot2 version).
}
\examples{
model <- lm(data = mtcars, formula = mpg ~ wt + gear)
ggdiagnose(model, bins = NROW(mtcars), se = FALSE, freqpct = TRUE)
}
\seealso{
\url{https://github.com/robertschnitman/diagnoser}
}
