% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeldf.R
\name{modeldf}
\alias{modeldf}
\title{Transform lm, glm, or nls object into a data frame with margin of error and confidence intervals.}
\usage{
modeldf(model, conf = 0.95)
}
\arguments{
\item{model}{An lm, glm, or nls model object.}

\item{conf}{A confidence level, 0-1.}
}
\value{
A data frame.
}
\description{
Transform lm, glm, or nls object into a data frame with margin of error and confidence intervals.
}
\examples{
model.glm <- glm(data = mtcars, formula = am ~ mpg + gear, family = binomial(link = 'logit'))
modeldf(model = model.glm, conf = 0.90)
}
\seealso{
\url{https://github.com/robertschnitman/diagnoser}
}
