% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_df_diag.R
\name{names.tbl_df_diag}
\alias{names.tbl_df_diag}
\title{names.tbl_df_diag}
\usage{
\method{names}{tbl_df_diag}(x)
}
\arguments{
\item{x}{An object of \code{tbl_df_diag}.}
}
\description{
names.tbl_df_diag
}
