% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_df_diag.R
\name{new_tibble_diag}
\alias{new_tibble_diag}
\title{new_tibble_diag}
\usage{
new_tibble_diag(x, xg, model)
}
\arguments{
\item{x}{tibble object (level = 1)}

\item{xg}{tibble object of higher levels}

\item{model}{model object}
}
\description{
Create a object of tbl_df_diag, subclass of tibble
}
