% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_df_diagnoser.R
\name{new_tibble_diagnoser}
\alias{new_tibble_diagnoser}
\title{Create a object of tbl_df_diagnoser, subclass of tibble}
\usage{
new_tibble_diagnoser(
  x,
  ks_test = NULL,
  levene_test = NULL,
  id_influential = NULL
)
}
\arguments{
\item{x}{tibble object}

\item{ks_test}{result of ks.test}

\item{levene_test}{result of leveneTest}

\item{id_influential}{row ID at which Cook's distance > 1}
}
\description{
Create a object of tbl_df_diagnoser, subclass of tibble
}
