% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnosis.R
\name{p2lt}
\alias{p2lt}
\title{Conversion from P-value to expression}
\usage{
p2lt(p)
}
\arguments{
\item{p}{P-value}
}
\description{
Conversion from P-value to expression
}
\examples{
p <- c(1e-05, 0.002, 0.03, 0.4)
p2lt(p)
}
