% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnosis.R
\name{p2star}
\alias{p2star}
\title{Conversion from P-value to stars}
\usage{
p2star(p)
}
\arguments{
\item{p}{P-value}
}
\description{
Conversion from P-value to stars
}
\examples{
p <- c(1e-05, 0.002, 0.03, 0.4)
p2star(p)
}
