% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnosis.R
\name{plot_added_variable}
\alias{plot_added_variable}
\title{plot_added_variable}
\usage{
plot_added_variable(object, mapping, ...)
}
\arguments{
\item{object}{An object of \code{tbl_df_diag}.}

\item{mapping}{An aesthetic mapping object \code{aes()} of \code{ggplot2} with a \code{x} argument.}

\item{...}{currently not to be used.}
}
\description{
Added-variable plot
}
\note{
To generate an added-variable plot.
}
