% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnosis.R
\name{plot_component_residual}
\alias{plot_component_residual}
\title{plot_component_residual}
\usage{
plot_component_residual(object, mapping, ...)
}
\arguments{
\item{object}{An object of \code{tbl_df_diag}.}

\item{mapping}{An aesthetic mapping object \code{aes()} of \code{ggplot2} with a \code{x} argument.}

\item{...}{currently not to be used.}
}
\description{
Component+residual plot
}
\note{
To generate a component-plus-residual plot.
}
