% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnosis.R
\name{plot_cooksd}
\alias{plot_cooksd}
\title{plot_cooksd}
\usage{
plot_cooksd(object, mapping, ...)
}
\arguments{
\item{object}{An object of \code{tbl_df_diag}.}

\item{mapping}{An aesthetic mapping object \code{aes()} of \code{ggplot2}.}

\item{...}{currently not to be used.}
}
\description{
Plot of Cook's distance
}
\note{
To visualize the Cook's distance.
}
