% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnosis.R
\name{plot_diagnosis.lm}
\alias{plot_diagnosis.lm}
\title{Diagnosis plot fo lm object}
\usage{
\method{plot_diagnosis}{lm}(model, ...)
}
\arguments{
\item{model}{lm object}

\item{...}{Not to be used.}
}
\description{
Diagnosis plot fo lm object
}
\note{
It produces four plots: Residual vs Fitted, Normal QQ, Scale-Location, and Cook's distance.
}
\examples{
fit <- lm(weight ~ height, women)
plot_diagnosis(fit)
}
