% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnosis.R
\name{plot_diagnosis.lme}
\alias{plot_diagnosis.lme}
\title{Diagnosis plot fo lme object}
\usage{
\method{plot_diagnosis}{lme}(model, ...)
}
\arguments{
\item{model}{lme object}

\item{...}{Not to be used.}
}
\description{
Diagnosis plot fo lme object
}
\note{
four plots of Residual vs Fitted, Normal QQ, Scale-Location, and Cook's distance
}
\examples{
library(nlme)
fit <- lme(circumference ~ age, random = ~ 1|Tree, Orange)
plot_diagnosis(fit)
}
\author{
ISHIKURA Kiwamu
}
\keyword{function}
