% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnosis.R
\name{plot_diagnosis.lmerMod}
\alias{plot_diagnosis.lmerMod}
\title{Diagnosis plot fo lmerMod object}
\usage{
\method{plot_diagnosis}{lmerMod}(model, ...)
}
\arguments{
\item{model}{lmerMod object}

\item{...}{Not to be used.}
}
\description{
Diagnosis plot fo lmerMod object
}
\note{
It produces four plots: Residual vs Fitted, Normal QQ, Scale-Location, and Cook's distance.
}
\examples{
library(lme4)
fit <- lmer(circumference ~ age + (1|Tree), Orange)
plot_diagnosis(fit)
}
