% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnosis.R
\name{plot_diagnosis}
\alias{plot_diagnosis}
\title{Generic function of plot_diagnosis}
\usage{
plot_diagnosis(model, ...)
}
\arguments{
\item{model}{model object}

\item{...}{additional arguments}
}
\description{
Generic function of plot_diagnosis
}
\note{
This is a generic function to make the diagnosis plot of the model.
}
\seealso{
\code{\link{plot}}
}
\author{
ISHIKURA Kiwamu
}
\keyword{function}
