% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnosis.R
\name{plot_marginal_model}
\alias{plot_marginal_model}
\title{plot_marginal_model}
\usage{
plot_marginal_model(object, mapping, ...)
}
\arguments{
\item{object}{An object of \code{tbl_df_diag}.}

\item{mapping}{An aesthetic mapping object \code{aes()} of \code{ggplot2} with a \code{x} argument.}

\item{...}{currently not to be used.}
}
\description{
Marginal-model plot
}
\note{
To generate a marginal-model plot.
}
