% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnosis.R
\name{plot_qq}
\alias{plot_qq}
\title{plot_qq}
\usage{
plot_qq(object, mapping, level = 1, ...)
}
\arguments{
\item{object}{An object of \code{tbl_df_diag}.}

\item{mapping}{An aesthetic mapping object \code{aes()} of \code{ggplot2}.}

\item{level}{A character of level for mixed-effects models (default = 1). See \code{\link[HLMdiag]{hlm_influence}}.}

\item{...}{currently not to be used.}
}
\description{
QQ plot
}
\note{
To generate a QQ plot.
}
