% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnosis.R
\name{plot_scale_location}
\alias{plot_scale_location}
\title{plot_scale_location}
\usage{
plot_scale_location(object, mapping, ...)
}
\arguments{
\item{object}{An object of \code{tbl_df_diag}.}

\item{mapping}{An aesthetic mapping object \code{aes()} of \code{ggplot2}.}

\item{...}{currently not to be used.}
}
\description{
Scale-location plot
}
\note{
To generate a Scale-Location plot.
}
