% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Common model statistics in matrix form.}
\usage{
validate(model)
}
\arguments{
\item{model}{An lm or glm object.}
}
\value{
Vector (column).
}
\description{
Common model statistics in matrix form.
}
\examples{
model.lm <- lm(data = mtcars, formula = mpg ~ wt + gear)
validate(model.lm)

model.glm <- glm(data = mtcars, am ~ mpg + wt, family = binomial(link = 'logit'))
validate(model.glm)
}
\seealso{
\url{https://github.com/robertschnitman/diagnoser}
}
