% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfreqdf.R
\name{wfreqdf}
\alias{wfreqdf}
\title{Convert raw text into a word-frequency data frame.}
\usage{
wfreqdf(filename, stopwords)
}
\arguments{
\item{filename}{A raw text document, preferably in a .txt file.}

\item{stopwords}{Specification for stopwords.}
}
\value{
A data frame.
}
\description{
Convert raw text into a word-frequency data frame.
}
\examples{
wfreqdf('nonsense.txt', 'english')
}
\seealso{
\url{https://github.com/robertschnitman/diagnoser}
}
