% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdiagnose.R
\name{cdiagnose}
\alias{cdiagnose}
\title{Graphically diagnose model residuals ("classic" version with ggplot2 graphics).}
\usage{
cdiagnose(model)
}
\arguments{
\item{model}{An lm or glm object.}
}
\value{
2x2 charts similar to plot(model.lm).
}
\description{
Graphically diagnose model residuals ("classic" version with ggplot2 graphics).
}
\examples{
model.lm <- lm(data = mtcars, formula = mpg ~ wt + gear)
cdiagnose(model.lm)
}
\seealso{
\url{https://github.com/robertschnitman/diagnoser}
Raju Rimal (inspiration for Residuals vs. Leverage): \url{https://rpubs.com/therimalaya/43190}
}
