% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnoseR.R
\name{comp_alg}
\alias{comp_alg}
\title{Compare Algoritmos de Amprendizado de Máquina.}
\usage{
comp_alg(
  data,
  target,
  list_alg = c("rpart", "nnet", "svmLinear", "rf", "LogitBoost", "knn"),
  train_val = 0.75,
  cv_folds = 5,
  seed = 123,
  number = 5,
  repeats = 10,
  verbose = TRUE
)
}
\description{
Esta função compara multiplos algoritmos de aprendizado de máquina.
}
