% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{diagnose.aov}
\alias{diagnose.aov}
\title{Diagnose a aov model}
\usage{
\method{diagnose}{aov}(model)
}
\arguments{
\item{model}{aov object}
}
\description{
Diagnose a aov model
}
\note{
To diagnose the aov model
}
