% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{diagnose.lm}
\alias{diagnose.lm}
\title{Diagnose a lm model}
\usage{
\method{diagnose}{lm}(model)
}
\arguments{
\item{model}{\code{\link[stats]{lm}} object.}
}
\value{
An object of \code{tbl_df_diag}.
}
\description{
Diagnose a lm model
}
\note{
To diagnose the \code{\link[stats]{lm}} model.
}
\examples{
lm_rock <- lm(perm ~ area + peri + shape, rock)
(diag_rock <- diagnose(aov_rock))
}
