% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{diagnose.lme}
\alias{diagnose.lme}
\title{Diagnose a lme model}
\usage{
\method{diagnose}{lme}(model)
}
\arguments{
\item{model}{lme object}
}
\description{
Diagnose a lme model
}
\note{
To diagnose the lme model
}
