% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{diagnose.lmerMod}
\alias{diagnose.lmerMod}
\title{Diagnose a lmerMod model}
\usage{
\method{diagnose}{lmerMod}(model)
}
\arguments{
\item{model}{lmerMod object}
}
\description{
Diagnose a lmerMod model
}
\note{
To diagnose the lmerMod model
}
