% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{diagnose.lmerMod}
\alias{diagnose.lmerMod}
\title{diagnose.lmerMod}
\usage{
\method{diagnose}{lmerMod}(model, ...)
}
\arguments{
\item{model}{\code{lmerMod} object.}

\item{...}{Currently not to be used.}
}
\value{
An object of \code{tbl_df_diag}.
}
\description{
To generate a diagnosis of a \code{lmerMod} model.
}
\note{
To diagnose the \code{lmerMod} model.
}
\examples{
library(nlme)
library(lme4)
lmer_orthodont <- lmer(distance ~ age + Sex + (age + Sex | Subject),
                       data = Orthodont)
(diag_orthodont <- diagnose(lmer_orthodont))
}
