% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{diagnose}
\alias{diagnose}
\title{Graphically diagnose model residuals.}
\usage{
diagnose(model, fit_type = "response", residual_type = "response")
}
\arguments{
\item{model}{An lm or glm object.}

\item{fit_type}{String. Default is "response". Type of fitted values to use based on options in predict().}

\item{residual_type}{String. Default is "response". Type of residuals values to use based on options in resid().}
}
\value{
2x2 charts similar to plot(model.lm): 2 scatter plots and 2 histograms of residuals and "residuals margin,"
which is the residuals as a percentage of the actual dependent variable values.
}
\description{
Graphically diagnose model residuals.
}
\examples{
model.lm <- lm(data = mtcars, formula = mpg ~ wt + gear)
diagnose(model.lm)
}
\seealso{
\url{https://github.com/robertschnitman/diagnoser}
}
