% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{diagnose}
\alias{diagnose}
\title{Graphically diagnose model residuals.}
\usage{
diagnose(model)
}
\arguments{
\item{model}{An lm or glm object.}
}
\value{
2x2 charts similar to plot(model.lm).
}
\description{
Graphically diagnose model residuals.
}
\examples{
model.lm <- lm(data = mtcars, formula = mpg ~ wt + gear)
diagnose(model.lm)
}
\seealso{
\url{https://github.com/robertschnitman/diagnoser}
}
