% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{diagnose}
\alias{diagnose}
\title{Graphically diagnose model residuals.}
\usage{
diagnose(model, fit_type = "response", residual_type = "response",
  point_color = "black", line_color = "black", pch = 1, lwd = 1)
}
\arguments{
\item{model}{An lm, glm, or nls object.}

\item{fit_type}{String. Default is "response". Type of fitted values to use based on options in predict().}

\item{residual_type}{String. Default is "response". Type of residuals values to use based on options in resid().}

\item{point_color}{Color of the points and fill color of the histograms. If left to defaults, fill of the histogram is white.}

\item{line_color}{Color of the smoothed line.}

\item{pch}{Point type. Same settings as Base R pch in plot().}

\item{lwd}{Width of the smoothed line.}
}
\value{
2x2 charts similar to plot(model.lm): 2 scatter plots and 2 histograms of residuals and "residuals margin,"
which is the residuals as a percentage of the actual dependent variable values.
}
\description{
Graphically diagnose model residuals.
}
\section{Warning}{

NLS objects will only work if "model = TRUE" is specified in the original NLS function.
}

\examples{

# OLS case
model.lm <- lm(data = mtcars, formula = mpg ~ wt + gear)
diagnose(model.lm) # defaults
diagnose(model.lm, point_color = '#00BFC4', line_color = '#F8766D', pch = 16, lwd = 2)

# NLS case
require(graphics)
DNase1    <- subset(DNase, Run == 1)
fm1DNase1 <- nls(density ~ SSlogis(log(conc), Asym, xmid, scal), DNase1, model = TRUE)
diagnose(fm1DNase1)

}
\seealso{
\url{https://github.com/robertschnitman/diagnoser}
}
