% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitres.R
\name{fitres}
\alias{fitres}
\title{Create matrix of model fit, residuals, and residuals margin (residuals as a proportion of the actual values).}
\usage{
fitres(model, type = "response")
}
\arguments{
\item{model}{An lm or glm object.}

\item{type}{String. Prediction type depends on whether the object is lm ('response', 'terms') or glm ('link', 'response', 'terms'). (See ?predict.lm and ?predict.glm for details).}

\item{data}{A data frame. If not specified, then the original data is used (i.e. model.frame(model)).}
}
\value{
A matrix.
}
\description{
Create matrix of model fit, residuals, and residuals margin (residuals as a proportion of the actual values).
}
\details{
fitres() creates a matrix of the fitted values, residuals, and residuals as a proportion (percent) of the actual dependent variable values based on an OLS model or GLM.
}
\examples{
model.lm <- lm(data = mtcars, formula = mpg ~ wt + gear)
fitres(model.lm, type = 'response')
}
\seealso{
\url{https://github.com/robertschnitman/diagnoser}
}
