% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitres.R
\name{fitres}
\alias{fitres}
\title{Create matrix of model fit, residuals, and residuals % (proportion).}
\usage{
fitres(model, data = NULL, fit_type = 'response', residual_type = 'response')
}
\arguments{
\item{model}{An lm, glm, or nls(model = TRUE) object.}

\item{data}{Dataframe. If defined, column-wise binds predictions/residuals to dataframe.}

\item{fit_type}{String. Prediction type. See ?predict for details.}

\item{residual_type}{String. See ?resid for details.}
}
\value{
A matrix OR dataframe (if data is not null).
}
\description{
Create matrix of model fit, residuals, and residuals % (proportion).
}
\details{
fitres() creates a matrix of the fitted values, residuals, and residuals as a proportion (percent) of the actual dependent variable values based on an OLS model or GLM. If the data input is defined, then said matrix is column-wise binded to the specified dataset.
}
\section{Warning}{

NLS objects will only work if "model = TRUE" is specified in the original NLS function.
}

\examples{
# OLS case
model.lm <- lm(data = mtcars, formula = mpg ~ wt + gear)
fitres(model.lm, fit_type = 'response')
fitres(model.lm, data = mtcars)

# NLS case
model.nls <- nls(Ozone ~ theta0 + Temp^theta1, airquality, model = TRUE)
fitres(model.nls)

}
\seealso{
\url{https://github.com/robertschnitman/diagnoser}
}
