% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitres.R
\name{fitres}
\alias{fitres}
\title{Create matrix of model fit, residuals, and residuals as a proportion of the fitted values.}
\usage{
fitres(model)
}
\arguments{
\item{model}{An lm or glm object.}
}
\value{
A matrix.
}
\description{
Create matrix of model fit, residuals, and residuals as a proportion of the fitted values.
}
\examples{
model.lm <- lm(data = mtcars, formula = mpg ~ wt + gear)
fitres(model.lm)
}
