% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitresdf.R
\name{fitresdf}
\alias{fitresdf}
\title{Merge fitted values and residuals to the original data frame.}
\usage{
fitresdf(model, data, type = "response")
}
\arguments{
\item{model}{An lm or glm object.}

\item{data}{A data frame. If unspecified, then the original data is used (i.e. model.frame(model)).}

\item{type}{String. Prediction type depends on whether the object is lm ('response', 'terms') or glm ('link', 'response', 'terms'). (See ?predict.lm and ?predict.glm for details).}
}
\value{
A data frame.
}
\description{
Merge fitted values and residuals to the original data frame.
}
\examples{
model.lm <- lm(data = mtcars, formula = mpg ~ wt + gear)
fitresdf(data = mtcars, model = model.lm)

# A warning message displays when there are missing values in the dataset.
df       <- mtcars
df[1,1]  <- NA
model.lm <- lm(data = mtcars, formula = mpg ~ wt + gear)
fitresdf(model.lm, data = df)
}
\seealso{
\url{https://github.com/robertschnitman/diagnoser}
}
