% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitresdf.R
\name{fitresdf}
\alias{fitresdf}
\title{Append fitted values and residuals to a data frame.}
\usage{
fitresdf(data, model)
}
\arguments{
\item{data}{A data frame.}

\item{model}{An lm or glm object.}
}
\value{
A data frame.
}
\description{
Append fitted values and residuals to a data frame.
}
\examples{
model.lm <- lm(data = mtcars, formula = mpg ~ wt + gear)
head(fitresdf(data = mtcars, model = model.lm))
}
\seealso{
\url{https://github.com/robertschnitman/schnitr}
}
