% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_var_importance.R
\name{get_var_importance}
\alias{get_var_importance}
\title{Get Variable Importance from Trained Models}
\usage{
get_var_importance(diagnoseR_result)
}
\arguments{
\item{diagnoseR_result}{A list object of class `diagnoseR_result` from `comp_alg()`.}
}
\value{
A list where each element is a data frame of variable importances for an algorithm,
sorted from most to least important based on the best model's ranking.
}
\description{
This function extracts variable importance from the models trained by `comp_alg`.
It standardizes the output by ordering all variables according to their
importance in the best model (identified by `comp_alg`). This ensures that
all variables are present in the output for every model, facilitating direct comparisons.
}
\examples{
# First, run comp_alg to get a results object
iris_results <- comp_alg(data = iris, target = "Species", verbose = FALSE, seed = 123)

# Get importance from all models
var_imp_list <- get_var_importance(iris_results)
# Print importance for the 'rf' model
# Note the variables are ordered based on the best model's (svmLinear) importance
print(var_imp_list$rf)

}
