% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdiagnose.R
\name{ggdiagnose}
\alias{ggdiagnose}
\title{Graphically diagnose model residuals (ggplot2 version).}
\usage{
ggdiagnose(model, bins = 30)
}
\arguments{
\item{model}{An lm or glm object.}

\item{bins}{Number of bins to specify for histograms.}
}
\value{
2x2 charts similar to plot(model.lm).
}
\description{
Graphically diagnose model residuals (ggplot2 version).
}
\examples{
model.lm <- lm(data = mtcars, formula = mpg ~ wt + gear)
ggdiagnose(model.lm, bins = NROW(mtcars))
}
\seealso{
\url{https://github.com/robertschnitman/diagnoser}
}
