% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdiagnose.R
\name{ggdiagnose}
\alias{ggdiagnose}
\title{Graphically diagnose model residuals (ggplot2 version).}
\usage{
ggdiagnose(model, fit_type = "response", residual_type = "response",
  bins = 30, se = TRUE, freqpct = FALSE, alpha = 1)
}
\arguments{
\item{model}{An lm or glm object.}

\item{fit_type}{String. Default is "response". Type of fitted values to use based on options in predict().}

\item{residual_type}{String. Default is "response". Type of residuals values to use based on options in resid().}

\item{bins}{Number of bins to specify for histograms.}

\item{se}{Boolean. For overlaying shaded standard errors.}

\item{freqpct}{Boolean.}

\item{alpha}{Integer, [0, 1]. Points are more transparent the closer they are to 0. Only applies to scatter plots.}
}
\value{
2 scatter plots and 2 histograms of residuals and "residuals margin,"
which is the residuals as a percentage of the actual dependent variable values.
}
\description{
Graphically diagnose model residuals (ggplot2 version).
}
\section{Warning}{

NLS objects will only work if "model = TRUE" is specified in the original NLS function.
}

\examples{

# OLS case
model <- lm(data = mtcars, formula = mpg ~ wt + gear)
ggdiagnose(model, bins = NROW(mtcars), se = FALSE, freqpct = TRUE)

# NLS case
require(graphics)
DNase1    <- subset(DNase, Run == 1)
fm1DNase1 <- nls(density ~ SSlogis(log(conc), Asym, xmid, scal), DNase1, model = TRUE)
ggdiagnose(fm1DNase1)

}
\seealso{
\url{https://github.com/robertschnitman/diagnoser}
}
