% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmdf.R
\name{glmdf}
\alias{glmdf}
\title{Transform glm() object into a data frame with margin of error and confidence intervals.}
\usage{
glmdf(model, conf = 95)
}
\arguments{
\item{model}{A glm object.}

\item{conf}{A confidence level. Either 90, 95, or 99.}
}
\value{
A data frame.
}
\description{
Transform glm() object into a data frame with margin of error and confidence intervals.
}
\examples{
model.glm <- glm(data = mtcars, formula = am ~ mpg + gear, family = binomial(link = 'logit'))
glmdf(model = model.glm, conf = 90)
}
\seealso{
\url{https://github.com/robertschnitman/schnitr}
}
