% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmdf.R
\name{lmdf}
\alias{lmdf}
\title{Transform lm() object into a data frame with margin of error and confidence intervals.}
\usage{
lmdf(model, conf = 95)
}
\arguments{
\item{model}{An lm object.}

\item{conf}{A confidence level. Either 90, 95, or 99.}
}
\value{
A data frame.
}
\description{
Transform lm() object into a data frame with margin of error and confidence intervals.
}
\examples{
model.lm <- lm(data = mtcars, formula = mpg ~ wt + gear)
lmdf(model = model.lm, conf = 90)
}
