% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnosis.R
\name{autoplot.tbl_df_diag}
\alias{autoplot.tbl_df_diag}
\alias{plot.tbl_df_diag}
\title{autoplot.tbl_df_diag}
\usage{
\method{autoplot}{tbl_df_diag}(object, type = "mf", mapping = aes(), ...)

\method{plot}{tbl_df_diag}(x, type = "mf", mapping = aes(), ...)
}
\arguments{
\item{object}{An object of \code{tbl_df_diag}.}

\item{type}{A string. See details.}

\item{mapping}{An aesthetic mapping object \code{aes()} of \code{ggplot2}.}

\item{...}{currently not to be used.}

\item{x}{An object of \code{tbl_df_diag}.}
}
\description{
A generic function to plot an object of \code{tbl_df_diag} in various types.
}
\details{
The argument \code{type} can be specified as below:
* \code{mf} (\code{measured-fitted}): Measured vs fitted values (default)
* \code{rf} (\code{resid-fitted}): Studentized residual vs fitted values
* \code{qq}: QQ norm plot
* \code{sl} (\code{scale-location}): Scale-Location plot
* \code{cook} (\code{cooks-distance}): Cook's distance
* \code{mm} (\code{marginal-model}): Marginal-model plot
* \code{av} (\code{added-variable}): Added-variable plot (only for \code{lm} and \code{aov} object)
* \code{cr} (\code{component-plus-residual}): Component-plus residual plot (only for \code{lm} and \code{aov} object)
}
\examples{
library(ggplot2)
library(nlme)
lm_od <- lm(distance ~ age + Sex, Orthodont)
lme_od <- lme(distance ~ age + Sex, Orthodont, random = ~ age | Subject)
diag_lm_od <- diagnose(lm_od)
diag_lme_od <- diagnose(lme_od)

# measured-fitted plot
plot(diag_lme_od)

# residual-fitted plot
plot(diag_lme_od, type = "rf")

# residual plot of Subject in different Sex
plot(diag_lme_od, type = "rf", aes(x = Subject)) + facet_wrap(~ Sex, scales = "free_x")

# QQ plot
plot(diag_lme_od, type = "qq")

# QQ plot for random intercept
plot(diag_lme_od, type = "qq", aes(sample = .ranef.intercept), level = "Subject")

# QQ plot for random slope
plot(diag_lme_od, type = "qq", aes(sample = .ranef.age), level = "Subject")

# scale-location plot
plot(diag_lme_od, type = "sl")

# Cook's distance
plot(diag_lme_od, type = "cook")

# marginal-model plot
plot(diag_lme_od, type = "mm", mapping = aes(x = age))

# added-variable plot (only for lm and aov)
plot(diag_lm_od, type = "av", mapping = aes(x = age))

# component-plus-residual plot (only for lm and aov)
plot(diag_lm_od, type = "cr", mapping = aes(x = age))
}
