% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnosis.R
\name{plot_added_variable}
\alias{plot_added_variable}
\title{plot_added_variable}
\usage{
plot_added_variable(object, mapping, ...)
}
\arguments{
\item{object}{An object of \code{tbl_df_diag}.}

\item{mapping}{An \code{\link[ggplot2]{aes()}} object with a \code{x} argument.}
}
\description{
Added-variable plot
}
\note{
To generate an added-variable plot.
}
