% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnosis.R
\name{plot_component_residual}
\alias{plot_component_residual}
\title{plot_component_residual}
\usage{
plot_component_residual(object, mapping, ...)
}
\arguments{
\item{object}{An object of \code{tbl_df_diag}.}

\item{mapping}{An \code{\link[ggplot2]{aes()}} object with a \code{x} argument.}
}
\description{
plot_component_residual
}
\note{
To generate a component-plus-residual plot.
}
