% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnosis.R
\name{plot_marginal_model}
\alias{plot_marginal_model}
\title{plot_marginal_model}
\usage{
plot_marginal_model(object, mapping, ...)
}
\arguments{
\item{object}{An object of \code{tbl_df_diag}.}

\item{mapping}{An \code{\link[ggplot2]{aes()}} object with a \code{x} argument.}
}
\description{
plot_marginal_model
}
\note{
To generate a marginal-model plot.
}
