% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnosis.R
\name{plot_measured_fitted}
\alias{plot_measured_fitted}
\title{plot_measured_fitted}
\usage{
plot_measured_fitted(object, mapping, ...)
}
\arguments{
\item{object}{An object of \code{tbl_df_diag}.}

\item{mapping}{An \code{\link[ggplot2]{aes()}} object.}
}
\description{
plot_measured_fitted
}
\note{
To generate a scatter plot of measured vs fitted values.
}
