% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnosis.R
\name{plot_resid_fitted}
\alias{plot_resid_fitted}
\title{plot_resid_fitted}
\usage{
plot_resid_fitted(object, mapping, ...)
}
\arguments{
\item{object}{An object of \code{tbl_df_diag}.}

\item{mapping}{An aesthetic mapping object \code{aes()} of \code{ggplot2}.}

\item{...}{currently not to be used.}
}
\description{
Plot of studentized residuals vs fitted values.
}
\note{
To generate a scatter plot of fitted values vs studentized residuals.
}
