% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnosis.R
\name{plot_scale_location}
\alias{plot_scale_location}
\title{plot_scale_location}
\usage{
plot_scale_location(object, mapping, ...)
}
\arguments{
\item{object}{An object of \code{tbl_df_diag}.}

\item{mapping}{An \code{\link[ggplot2]{aes()}} object.}
}
\description{
Scale-location plot
}
\note{
To generate a Scale-Location plot.
}
