% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnoseR.R
\name{plot_var_importance}
\alias{plot_var_importance}
\title{Plot Variable Importance}
\usage{
plot_var_importance(importance_list, model_name, top_n = 15)
}
\arguments{
\item{importance_list}{The list of variable importances from `get_var_importance()`.}

\item{model_name}{The name of the model to plot (e.g., "rf", "LogitBoost").}

\item{top_n}{The number of top variables to display in the plot.}
}
\value{
A `ggplot` object showing the variable importance plot.
}
\description{
Creates a bar plot of variable importance for a specified model.
}
