% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_var_importance.R
\name{plot_var_importance}
\alias{plot_var_importance}
\title{Plot Variable Importance}
\usage{
plot_var_importance(importance_list, model_name, top_n = 15)
}
\arguments{
\item{importance_list}{The list of variable importances from `get_var_importance()`.}

\item{model_name}{The name of the model to plot (e.g., "rf", "LogitBoost").}

\item{top_n}{The number of top variables to display in the plot.}
}
\value{
A `ggplot` object showing the variable importance plot.
}
\description{
Creates a bar plot of variable importance for a specified model.
}
\examples{
# First, run comp_alg to get a results object
iris_results <- comp_alg(data = iris, target = "Species", verbose = FALSE, seed = 123)

# Get importance from the single best model trained on the full data
var_imp_full <- get_var_importance(iris_results, model_source = "full")

# The name of the best model is stored in the results object
best_model_name <- iris_results$best_model

# Plot the variable importance for the best model
if (length(var_imp_full) > 0) {
  plot_var_importance(var_imp_full, model_name = best_model_name)
}

}
