% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnoseR.R
\name{save_all_var_plots}
\alias{save_all_var_plots}
\title{Save All Variable Importance Plots}
\usage{
save_all_var_plots(
  importance_list,
  output_dir = "variable_importance_plots",
  top_n = 15,
  width = 8,
  height = 6,
  dpi = 300
)
}
\arguments{
\item{importance_list}{The list of variable importances from `get_var_importance()`.}

\item{output_dir}{The directory where the plot files will be saved. If it doesn't exist, it will be created.}

\item{top_n}{The number of top variables to display in each plot.}

\item{width}{The width of the saved plot in inches.}

\item{height}{The height of the saved plot in inches.}

\item{dpi}{The resolution for the saved plot.}
}
\value{
Invisibly returns a character vector of the saved file paths.
}
\description{
Generates and saves variable importance plots for all models to a specified directory.
}
