% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Common model statistics in a vector.}
\usage{
validate(model)
}
\arguments{
\item{model}{An lm or glm object.}
}
\value{
Vector (column). Includes F-statistic, R-squared, RMSE, and others.
}
\description{
Common model statistics in a vector.
}
\details{
The broom library's glance() had a vague label for the F statistic (simply "statistic") and lacked the pseudo R-squared, which is commonly based on McFadden's version (i.e. 1 - (residual deviance / null deviance)).
While the same function is friendly for data frames, it's wide form is cumbersome for quickly ascertaining model validity. Thus, validate() produces similar output as a column vector. Those who wish to have the values in broom's format can always transpose the vector.
}
\examples{
model.lm <- lm(data = mtcars, formula = mpg ~ wt + gear)
validate(model.lm)

model.glm <- glm(data = mtcars, am ~ mpg + wt, family = binomial(link = 'logit'))
validate(model.glm)
}
\seealso{
\url{https://github.com/robertschnitman/diagnoser}
}
